simple wrapper for polkadot api
